Basic charts for the histogram and sentiment tools
This commit adds the charts based on the query results. I included
gwt-visualization.jar in this submission. You don't need to add it to
the lib directory but you do need to add it to the build path. The code
is kind of hacky at the moment but it works.